requirements.txt
streamlit
requests
pandas
matplotlib
google-generativeai
